pyproject.toml
[project]
name = "hivewiki-builder"
version = "0.1.0"
description = "Content processing and wiki generation service for HiveWiki"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "anthropic>=0.49.0",
    "feedparser>=6.0.12",
    "ipykernel>=7.2.0",
    "ipython>=9.13.0",
    "knowledge-storm>=1.1.1",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
]

[dependency-groups]
dev = ["pre-commit>=4.5.1"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
